Reorganize use cases folder and rename classes
