+++ tests/use_cases/play/test_artist_request_objects.py
