+++ tests/use_cases/play/test_music_use_case.py
@@ -3,7 +3,7 @@ from mock import mock
 
 from snipssonos.entities.device import Device
 
-from snipssonos.use_cases.play_music import PlayMusicUseCase
+from snipssonos.use_cases.play.music import PlayMusicUseCase
 from snipssonos.use_cases.request_objects import PlayMusicRequestFactory
 
 from snipssonos.exceptions import NoReachableDeviceException
